requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.1
python-dotenv>=1.0.1
google-generativeai>=0.8.0
openai>=1.14.0
qdrant-client>=1.8.0
Pillow>=10.2.0
python-multipart>=0.0.9
httpx>=0.27.0
pytesseract>=0.3.10
deep-translator>=1.11.0
opencv-python>=4.9.0
